+++ core/src/server/delivery/RequestHandler.h
@@ -34,7 +34,7 @@ class RequestHandler {
 
     Status
     CreateTable(const std::shared_ptr<Context>& context, const std::string& table_name, int64_t dimension,
-                int32_t index_file_size, int32_t metric_type);
+                int64_t index_file_size, int64_t metric_type);
 
     Status
     HasTable(const std::shared_ptr<Context>& context, const std::string& table_name, bool& has_table);
@@ -43,8 +43,8 @@ class RequestHandler {
     DropTable(const std::shared_ptr<Context>& context, const std::string& table_name);
 
     Status
-    CreateIndex(const std::shared_ptr<Context>& context, const std::string& table_name, int32_t index_type,
-                int32_t nlist);
+    CreateIndex(const std::shared_ptr<Context>& context, const std::string& table_name, int64_t index_type,
+                int64_t nlist);
 
     Status
     Insert(const std::shared_ptr<Context>& context, const std::string& table_name, int64_t record_size,
